added a restart to program and cleaned up
cleared the buttons of the previous frames and added a final results frame which hides the widgets after finishing the code using 'global'. This makes the program cleaner. However, there i a problem with the padding, causing the program to go down every time I press the restart button. I am also looking forward to adding multiple buttons for the program.