Created operations quiz corresponding to each button
Using the parameter that I have provided in each of the button commands, I was able to add an if statement on the parameter which corresponds to the desired operation. The next steps to finishing the program is adding a restart button on top of the main menu and fixing the padding issue which seems to push the  quiz down every time the main menu button is pressed.