Major design changes, fixed widget ordering
I fixed the widget ordering of the program during the framerun() container since the question number heading was at the bottom. I reversed the pack order of each widget and used side='bottom' to put the heading at the top. I also added colour to the program as well as enlarging the buttons and text for better readability. The main menu buttons now take up the entire menu.